// data/github/current/ui/packages/issues-react/components/dashboard/__tests__/__generated__/DashboardSearchTestQuery.graphql.ts
export type DashboardSearchTestQuery$variables = {
  first?: number | null | undefined;
  labelPageSize?: number | null | undefined;
  query?: string | null | undefined;
  skip?: number | null | undefined;
};
export type DashboardSearchTestQuery$data = {
  readonly " $fragmentSpreads": FragmentRefs<"DashboardSearchFragment">;
};
export type DashboardSearchTestQuery = {
  response: DashboardSearchTestQuery$data;
  variables: DashboardSearchTestQuery$variables;
};